database_run.py completed
Update operation added. Program will first delete any restaurants with the name 'McDonalds' if they exist (needed if lotsofmenus.py previously executed). Then it will insert 2 resetaurants and 1 menu item. It will then proceed to query the data and filter the database for a restaurant with the name 'McDonalds' and change it to 'Panera Bread'. Then it will delete all values in both tables so lotsofmenus.py can be run with an empty db.