Interactive web application now allows CRUD operations for our restaurantmenu database
The homepage is located at localhost:8080/restaurants which lists all of the restaurants in the database. On the homepage, there is a link to create a new restaurant that will redirect you to the link 'localhost:8080/restaurants/new'
and allow you to specify a new restaurant name which is added to the db. Underneath
each restaurant on the homepage there is a link to 'Edit' the name of a restaurant, that will redirect you to 'localhost:8080/restaurants/[id]/edit' which will find the name of the restaurant you want to edit and update the db with the new name. Beneath each edit link there is a 'Delete' link that brings you to 'localhost:8080/restaurants/[id]/delete' and likewise gives you the option to delete the restaurant from the db or cancel the action.

The program specifies that the 'database_setup' file is inside the IntroToSQLAlchemy folder.This is intended to maintain folder hierarchy on Github and keep each lesson separate. Python only allows importing of files from the current directory. However, we can work around this by appending the subdirectory to the current path and giving Python access to the files within. The database is also inside this folder and its name must be hardcoded as '/IntroToSQLAlchemy/restaurantmenu.db'.