Creating our web application using Flask framework
Now we are going to implement the web application we made in 'webserver.py' using the Flask framework to expedite the process of writing a web server app that can receive and respond to user requests and to cut down on the amount of code we have to write.

We will connect to the db using SQLAlchemy, incorporate templates to separate our python code from our HTML code, use forms to capture data from users and use message flashing to notify user each time they successfully change information in the database.