Insert and Read Operations added to 'project.py' application
Our program lists the menu for a specific restaurant given the restaurant id. Each restaurant has it's own menu list that is reachable at localhost:5000/restaurants/[restuarant_id]. For every menu item: a user can create a new item, edit an existing item, or delete an existing item. When the user clicks on the available links, the program can reroute to the appropriate path.

-'Edit' routes to localhost:5000/restaurants/[restaurant_id]/[menu_id]/edit
-'Delete' routes to localhost:5000/restaurants/[restaurant_id]/[menu_id]/delete
-'' routes to localhost:5000/restaurants/[restaurant_id]/[menu_id]/edit

So far, templates for creating a new menu item and listing all menu items per restaurant have been created, and will be included in a subsequent commit. Templates for editing a menu item and deleting an item are soon to follow.