Files added.
'login.html' is used for logging in a user. User can login with google or facebook by clicking on corresponding buttons. They will then be redirected to homepage.

'publicmenu.html' and 'publicrestaurants.html' are displayed when a user has not logged in. A user can still browse restaurants and menu page but they cannot make changes, and the options for doing so are not available.